CapGen/requirements.txt
torch==1.9.0
transformers==4.11.3
streamlit==1.5.0
Pillow==8.3.2
